#1407: Remove repetead code and move to func package
