+++ src/test/java/org/cactoos/experimental/TimedTest.java
@@ -30,8 +30,8 @@ import java.util.concurrent.CancellationException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import org.cactoos.RepeatedCallable;
 import org.cactoos.func.CallableOf;
+import org.cactoos.func.RepeatedCallable;
 import org.cactoos.scalar.ScalarOf;
 import org.hamcrest.core.IsNull;
 import org.junit.Test;
